fix(account): resolve custom profile field metadata from full catalog
Expose customProfileFieldCatalog on well-known so Account Center can show
required labels and field types for profile fields outside the sign-up subset.
