+++ packages/core/src/libraries/sign-in-experience/index.ts
@@ -336,6 +336,7 @@ export const createSignInExperienceLibrary = (
       googleOneTap: getGoogleOneTap(),
       captchaConfig: await getCaptchaConfig(),
       customProfileFields: signUpCustomProfileFields,
+      customProfileFieldCatalog: customProfileFields,
     };
   };
